refactor to start() and stopContainer()
